// src/app/interfaces/informacion-general.ts
export interface InformacionGeneral {
    numeroAtencion: string;
    cedula: string;
    nombre: string;
    fechaNacimiento: Date;
    anosEscolaridad: number;
    fechaEvaluacion1: Date;
    edad: number;
    lateralidad: string;
    telefonoPaciente: string;
    telefonoAcompanante: string;
    emailContacto: string;
    ocupacionPrevia: string;
    ocupacionActual: string;
    ocupacionMayorTiempo: string;
    ciudad: string;
    lugarVivienda: string;
    informanteCuidador: string;
    parentesco: string;
    aseguradora: string;
    remite: string;
    longitudinal: boolean;
  }
  